Added CPU debugging support.
Added a controller class to handle SDL keyboard events.
Functions can be bound to a key down or key up event.

Added functions to pause, continue and step for a frame or single CPU instruction.

Added an array which could define breakpoints.
The emulator will automatically pause when PC equals any of the values
in the array.